# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor/shop:
    get:
      operationId: getVendorShop
      tags: [vendor]
      summary: Get the current user's own vendor shop info
      responses:
        "200":
          description: The caller's vendor shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorSummary"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
        "404":
          description: No vendor shop for this user
  /vendor/products:
    get:
      operationId: listVendorProducts
      tags: [vendor]
      summary: List the current user's own vendor products (all statuses)
      responses:
        "200":
          description: List of the vendor's products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
    post:
      operationId: createVendorProduct
      tags: [vendor]
      summary: Create a new product for the current user's own vendor shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorProductInput"
      responses:
        "200":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
        "401":
          description: Authentication required
        "403":
          description: Forbidden
  /vendor/products/{id}:
    patch:
      operationId: updateVendorProduct
      tags: [vendor]
      summary: Update one of the current user's own vendor products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteVendorProduct
      tags: [vendor]
      summary: Delete one of the current user's own vendor products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "401":
          description: Authentication required
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /vendor/analytics:
    get:
      operationId: getVendorAnalytics
      tags: [orders]
      summary: Sales/order/stock analytics for the current user's own vendor(s)
      responses:
        "200":
          description: Vendor analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorAnalytics"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
...
  /vendor/orders/{orderId}/fulfillment:
    patch:
      operationId: updateVendorOrderFulfillment
      tags: [orders]
      summary: Advance the fulfillment status for the caller's own vendor line items on an order
      description: |
        Only forward transitions are allowed (unfulfilled -> packed -> shipped
        -> delivered). Emits a `fulfillment.status_changed` domain event on
        each successful transition.
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFulfillmentStatusInput"
      responses:
        "200":
          description: Updated fulfillment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"
        "400":
          description: Invalid or out-of-order status transition
        "401":
          description: Authentication required
        "403":
          description: Forbidden
        "404":
          description: No fulfillment record for this vendor on this order
...
components:
  schemas:
...
    VendorSummary:
      type: object
      required: [id, slug, name, ratingAvgX100, ratingCount, badgeTopSeller, badgeFastShipper]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        ratingAvgX100: { type: integer }
        ratingCount: { type: integer }
        badgeTopSeller: { type: boolean }
        badgeFastShipper: { type: boolean }
        bio: { type: ["string", "null"] }
        logoImageUrl: { type: ["string", "null"] }
        bannerImageUrl: { type: ["string", "null"] }
    UpdateVendorProductInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        pricePesewas: { type: integer, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        isActive: { type: boolean }
    CreateVendorProductInput:
      type: object
      required: [categoryId, title, pricePesewas]
      properties:
        categoryId: { type: integer }
        title: { type: string, minLength: 1 }
        brand: { type: string }
        description: { type: string }
        pricePesewas: { type: integer, minimum: 0 }
        compareAtPesewas: { type: integer, minimum: 0 }
        stock: { type: integer, minimum: 0, default: 0 }
        tag: { type: string }
...
    Product:
      type: object
      required:
        [
          id,
          vendorId,
          categoryId,
          slug,
          title,
          pricePesewas,
          stock,
          ratingAvgX100,
          ratingCount,
          isActive,
        ]
      properties:
        id: { type: integer }
        vendorId: { type: integer }
        categoryId: { type: integer }
        slug: { type: string }
        title: { type: string }
        brand: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        pricePesewas: { type: integer }
        compareAtPesewas: { type: ["integer", "null"] }
        stock: { type: integer }
        tag: { type: ["string", "null"] }
        ratingAvgX100: { type: integer }
        ratingCount: { type: integer }
        isActive: { type: boolean }
        imageUrl: { type: ["string", "null"] }
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, fulfilled, cancelled]
...
    FulfillmentStatus:
      type: string
      enum: [unfulfilled, packed, shipped, delivered]
    Fulfillment:
      type: object
      required: [id, orderId, vendorId, status]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        vendorId: { type: integer }
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
        packedAt: { type: ["string", "null"], format: date-time }
        shippedAt: { type: ["string", "null"], format: date-time }
        deliveredAt: { type: ["string", "null"], format: date-time }
    UpdateFulfillmentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
...
    AnalyticsRevenuePoint:
      type: object
      required: [date, revenuePesewas, orderCount]
      properties:
        date: { type: string, format: date }
        revenuePesewas: { type: integer }
        orderCount: { type: integer }
    AnalyticsOrderStatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        count: { type: integer }
...
    VendorAnalyticsTopProduct:
      type: object
      required: [productId, title, unitsSold, revenuePesewas]
      properties:
        productId: { type: integer }
        title: { type: string }
        unitsSold: { type: integer }
        revenuePesewas: { type: integer }
    VendorAnalyticsLowStockProduct:
      type: object
      required: [productId, title, stock]
      properties:
        productId: { type: integer }
        title: { type: string }
        stock: { type: integer }
    VendorAnalytics:
      type: object
      required: [revenueSeries, orderStatusBreakdown, topProducts, lowStockProducts]
      properties:
        revenueSeries:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsRevenuePoint"
        orderStatusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsOrderStatusCount"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/VendorAnalyticsTopProduct"
        lowStockProducts:
          type: array
          items:
            $ref: "#/components/schemas/VendorAnalyticsLowStockProduct"
...